Help compiler detect an r-value
I don't know the C++ standard that well but this looks like a compiler
bug to me. The compiler failed to detect that "right" was an r-value and
did not allow us to call the move constructor of optional. Writing
`return std::move(right);` fixed it. But it's my understanding that you
never have to `move` a return value explicitly as that always happens
anyway.

In any case, we can just turn the code around get rid of `right`
altogether here.