Support {fmt}
This is much more convenient to use than the boost things (lexical_cast
and join). Also, this will be part of the C++20 standard so we likely
can drop {fmt} at some point and just use standard C++ again:
https://www.zverovich.net/2019/07/23/std-format-cpp20.html

Also, I hope that I can use this to drop some of the ostream<< hacks
that I need for cppyy to work. If not, I probably have to add
intervalxt::to_string() overloads as well.