Compute rational coefficients for several elements at a time
otherwise, the rational coefficients might be unrelated, e.g., when one
element is rational and another one non-rational, the rational vector of
the former might have length 1 while the other does not. So it is
unclear for the calling code how they relate without knowing too many
implementation details.

Also, this splits the Arithmetic into floor division and coefficients.

Also fixes #94.

Note that this is a breaking API change since the interface of Lengths
has changed. Also, the names of the sample includes have changed.